backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.35
cryptography>=41.0.7
PyPDF2>=3.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.0